Removed some more UI tests and a bit of refactoring in core.
