+++ src/main/resources/log4j2.xml
@@ -2,21 +2,22 @@
 <Configuration status="WARN">
   <Appenders>
     <Console name="Console" target="SYSTEM_OUT">
-      <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level - %msg%n" />
+      <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level - %msg%n"/>
     </Console>
-    <RollingFile 
-        name="FileLogger" 
-        fileName="logs/frameworkium.log"
-        filePattern="logs/frameworkium-%d{yyyy-MM-dd}.log"
-        append="true">
-      <PatternLayout pattern="%d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n" />
-      <SizeBasedTriggeringPolicy size="10MB" />
+    <RollingFile
+            name="FileLogger"
+            fileName="logs/frameworkium.log"
+            filePattern="logs/frameworkium-%d{yyyy-MM-dd}.log"
+            append="true">
+      <PatternLayout
+              pattern="%d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n"/>
+      <SizeBasedTriggeringPolicy size="10MB"/>
     </RollingFile>
   </Appenders>
   <Loggers>
     <Root level="trace">
-      <AppenderRef ref="Console" level="INFO" />
-      <AppenderRef ref="FileLogger" level="DEBUG" />
+      <AppenderRef ref="Console" level="INFO"/>
+      <AppenderRef ref="FileLogger" level="DEBUG"/>
     </Root>
   </Loggers>
 </Configuration>
