+++ src/test/java/com/frameworkium/integration/github/pages/web/SearchResultsPage.java
@@ -2,7 +2,6 @@ package com.frameworkium.integration.github.pages.web;
 
 import com.frameworkium.core.ui.annotations.Visible;
 import com.frameworkium.core.ui.pages.BasePage;
-import com.frameworkium.core.ui.pages.PageFactory;
 import com.frameworkium.integration.github.pages.web.components.HeaderComponent;
 import org.openqa.selenium.support.FindBy;
 import ru.yandex.qatools.allure.annotations.Step;
@@ -14,24 +13,14 @@ import java.util.stream.Collectors;
 
 public class SearchResultsPage extends BasePage<SearchResultsPage> {
 
-    @Name("Header")
     @Visible
+    @Name("Header")
     private HeaderComponent header;
 
-    @Name("Repository Links")
     @Visible
+    @Name("Repository Links")
     @FindBy(css = "h3 > a")
     private List<Link> repoLinks;
-
-    public HeaderComponent theHeader() {
-        return header;
-    }
-
-    @Step("Navigate to the Github homepage")
-    public static SearchResultsPage open() {
-        return PageFactory.newInstance(
-                SearchResultsPage.class, "http://github.com");
-    }
 
     @Step("Get the list of code repository names")
     public List<String> getRepoNames() {
